fix: swipeable & touchable working together
